# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/traspaso": {
      "post": {
        "tags": [
          "caja"
        ],
        "summary": "Traspaso",
        "operationId": "traspaso_api_traspaso_post",
        "parameters": [
          {
            "name": "sesion",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sesion"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraspasoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Saldo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/arqueo": {
      "post": {
        "tags": [
          "caja"
        ],
        "summary": "Arqueo",
        "operationId": "arqueo_api_arqueo_post",
        "parameters": [
          {
            "name": "sesion",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sesion"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArqueoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArqueoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rendicion": {
      "post": {
        "tags": [
          "caja"
        ],
        "summary": "Rendicion",
        "operationId": "rendicion_api_rendicion_post",
        "parameters": [
          {
            "name": "sesion",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sesion"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RendicionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RendicionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArqueoOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "efectivo_esperado": {
            "type": "integer",
            "title": "Efectivo Esperado"
          },
          "boletas_esperadas": {
            "type": "integer",
            "title": "Boletas Esperadas"
          },
          "diferencia_efectivo": {
            "type": "integer",
            "title": "Diferencia Efectivo"
          },
          "diferencia_boletas": {
            "type": "integer",
            "title": "Diferencia Boletas"
          },
          "estado": {
            "type": "string",
            "enum": [
              "cuadra",
              "con_diferencia",
              "explicada"
            ],
            "title": "Estado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "efectivo_esperado",
          "boletas_esperadas",
          "diferencia_efectivo",
          "diferencia_boletas",
          "estado"
        ],
        "title": "ArqueoOut"
      },
      "ArqueoRequest": {
        "properties": {
          "caja_id": {
            "type": "integer",
            "title": "Caja Id"
          },
          "turno_id": {
            "type": "integer",
            "title": "Turno Id"
          },
          "efectivo_contado": {
            "type": "integer",
            "exclusiveMaximum": 2147483648.0,
            "title": "Efectivo Contado"
          },
          "boletas_contadas": {
            "type": "integer",
            "exclusiveMaximum": 2147483648.0,
            "title": "Boletas Contadas"
          },
          "nota": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nota"
          }
        },
        "type": "object",
        "required": [
          "caja_id",
          "turno_id",
          "efectivo_contado",
          "boletas_contadas"
        ],
        "title": "ArqueoRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RendicionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "total_esperado": {
            "type": "integer",
            "title": "Total Esperado"
          },
          "total_contado": {
            "type": "integer",
            "title": "Total Contado"
          },
          "diferencia": {
            "type": "integer",
            "title": "Diferencia"
          },
          "cantidad_boletas": {
            "type": "integer",
            "title": "Cantidad Boletas"
          }
        },
        "type": "object",
        "required": [
          "id",
          "total_esperado",
          "total_contado",
          "diferencia",
          "cantidad_boletas"
        ],
        "title": "RendicionOut"
      },
      "RendicionRequest": {
        "properties": {
          "boletas_contadas": {
            "type": "integer",
            "exclusiveMaximum": 2147483648.0,
            "title": "Boletas Contadas"
          }
        },
        "type": "object",
        "required": [
          "boletas_contadas"
        ],
        "title": "RendicionRequest"
      },
...
      "Saldo": {
        "properties": {
          "efectivo": {
            "type": "integer",
            "title": "Efectivo"
          },
          "boletas": {
            "type": "integer",
            "title": "Boletas"
          }
        },
        "type": "object",
        "required": [
          "efectivo",
          "boletas"
        ],
        "title": "Saldo"
      },
...
      "TraspasoRequest": {
        "properties": {
          "caja_origen_id": {
            "type": "integer",
            "title": "Caja Origen Id"
          },
          "monto": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMaximum": 2147483648.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Monto"
          }
        },
        "type": "object",
        "required": [
          "caja_origen_id"
        ],
        "title": "TraspasoRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}